add chainconfigdb
