+++ internal/pkg/chainsdk/core/chaindata.go
@@ -62,7 +62,7 @@ func (d *ChainData) GetBlockForwardByReqTrx(trx *quorumpb.Trx, cipherKey string,
 
 	chain_log.Debugf("<%s> GetBlockForward block id: %s", trx.GroupId, reqBlockItem.BlockId)
 
-	subBlocks, err := d.dbmgr.GetSubBlock(reqBlockItem.BlockId, prefix...)
+	subBlocks, err := nodectx.GetNodeCtx().GetChainStorage().GetSubBlock(reqBlockItem.BlockId, prefix...)
 	return subBlocks, err
 }
 
@@ -106,7 +106,7 @@ func (d *ChainData) GetBlockBackwardByReqTrx(trx *quorumpb.Trx, cipherKey string
 		return nil, nil
 	}
 
-	isExist, err := d.dbmgr.IsBlockExist(reqBlockItem.BlockId, false, prefix...)
+	isExist, err := nodectx.GetNodeCtx().GetChainStorage().IsBlockExist(reqBlockItem.BlockId, false, prefix...)
 	if err != nil {
 		return nil, err
 	} else if !isExist {
@@ -118,12 +118,12 @@ func (d *ChainData) GetBlockBackwardByReqTrx(trx *quorumpb.Trx, cipherKey string
 		return nil, err
 	}
 
-	isParentExit, err := d.dbmgr.IsParentExist(block.PrevBlockId, false, prefix...)
+	isParentExit, err := nodectx.GetNodeCtx().GetChainStorage().IsParentExist(block.PrevBlockId, false, prefix...)
 	if err != nil {
 		return nil, err
 	}
 	if isParentExit {
-		parentBlock, err := d.dbmgr.GetParentBlock(reqBlockItem.BlockId, prefix...)
+		parentBlock, err := nodectx.GetNodeCtx().GetChainStorage().GetParentBlock(reqBlockItem.BlockId, prefix...)
 		return parentBlock, err
 	}
 	return nil, errors.New("Parent Block not exist")
@@ -190,7 +190,7 @@ func (d *ChainData) GetBlockForward(trx *quorumpb.Trx) (requester string, blocks
 	}
 
 	var subBlocks []*quorumpb.Block
-	subBlocks, err = d.dbmgr.GetSubBlock(reqBlockItem.BlockId, d.nodename)
+	subBlocks, err = nodectx.GetNodeCtx().GetChainStorage().GetSubBlock(reqBlockItem.BlockId, d.nodename)
 	if err != nil {
 		return "", nil, false, err
 	}
@@ -237,7 +237,7 @@ func (d *ChainData) GetBlockBackward(trx *quorumpb.Trx) (requester string, block
 		return reqBlockItem.UserId, nil, false, errors.New("insufficient privileges")
 	}
 
-	isExist, err := d.dbmgr.IsBlockExist(reqBlockItem.BlockId, false, d.nodename)
+	isExist, err := nodectx.GetNodeCtx().GetChainStorage().IsBlockExist(reqBlockItem.BlockId, false, d.nodename)
 	if err != nil {
 		return "", nil, false, err
 	} else if !isExist {
@@ -249,14 +249,14 @@ func (d *ChainData) GetBlockBackward(trx *quorumpb.Trx) (requester string, block
 		return "", nil, false, err
 	}
 
-	isParentExit, err := d.dbmgr.IsParentExist(blk.PrevBlockId, false, d.nodename)
+	isParentExit, err := nodectx.GetNodeCtx().GetChainStorage().IsParentExist(blk.PrevBlockId, false, d.nodename)
 	if err != nil {
 		return "", nil, false, err
 	}
 
 	if isParentExit {
 		chain_log.Debugf("<%s> send REQ_NEXT_BLOCK_RESP (BLOCK_IN_TRX)", d.groupId)
-		parentBlock, err := d.dbmgr.GetParentBlock(reqBlockItem.BlockId, d.nodename)
+		parentBlock, err := nodectx.GetNodeCtx().GetChainStorage().GetParentBlock(reqBlockItem.BlockId, d.nodename)
 		if err != nil {
 			return "", nil, false, err
 		}
