+++ internal/pkg/chainsdk/core/chainutils.go
@@ -78,7 +78,7 @@ func dfs(blocks []*quorumpb.Block, cache map[string]bool, result []string, noden
 		if _, ok := cache[block.BlockId]; !ok {
 			cache[block.BlockId] = true
 			result = append(result, block.BlockId)
-			subBlocks, err := nodectx.GetDbMgr().GetSubBlock(block.BlockId, nodename)
+			subBlocks, err := nodectx.GetNodeCtx().GetChainStorage().GetSubBlock(block.BlockId, nodename)
 			if err != nil {
 				return err
 			}
